rag-api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
psycopg2-binary==2.9.9
qdrant-client==1.7.0
pydantic==2.5.3
pydantic-settings==2.1.0
requests==2.31.0
python-multipart==0.0.6
